split proof conversion to file
