+++ crypto/bindings/conversion-core.ts
@@ -22,7 +22,6 @@ export {
   ConversionCore,
   ConversionCores,
   conversionCore,
-  conversionCorePerField,
   freeOnFinalize,
   wrap,
   unwrap,
