+++ crypto/bindings/conversion-oracles.ts
@@ -15,14 +15,7 @@ import {
   maybeFieldToRust,
 } from './conversion-base.js';
 
-export {
-  oraclesConversion,
-  oraclesConversionPerField,
-  OraclesConversion,
-  OraclesConversions,
-};
-
-// wasm types
+export { oraclesConversion };
 
 import { Field } from './kimchi-types.js';
 
@@ -36,9 +29,6 @@ type WasmClasses = {
   Oracles: typeof WasmFpOracles | typeof WasmFqOracles;
 };
 
-type OraclesConversion = ReturnType<typeof oraclesConversionPerField>;
-type OraclesConversions = ReturnType<typeof oraclesConversion>;
-
 function oraclesConversion(wasm: wasm) {
   const fp = oraclesConversionPerField({
     RandomOracles: wasm.WasmFpRandomOracles,
@@ -48,7 +38,6 @@ function oraclesConversion(wasm: wasm) {
     RandomOracles: wasm.WasmFqRandomOracles,
     Oracles: wasm.WasmFqOracles,
   });
-
   return { fp, fq };
 }
 
@@ -112,7 +101,7 @@ function oraclesConversionPerField({ RandomOracles, Oracles }: WasmClasses) {
     return mlRo;
   }
 
-  let self = {
+  return {
     oraclesToRust(oracles: Oracles): WasmOracles {
       let [, o, pEval, openingPrechallenges, digestBeforeEvaluations] = oracles;
       return new Oracles(
@@ -136,6 +125,4 @@ function oraclesConversionPerField({ RandomOracles, Oracles }: WasmClasses) {
       return mlOracles;
     },
   };
-
-  return self;
 }
