+++ crypto/bindings/conversion-verifier-index.ts
@@ -18,14 +18,7 @@ import {
   freeOnFinalize,
 } from './conversion-core.js';
 
-export {
-  verifierIndexConversion,
-  verifierIndexConversionPerField,
-  VerifierIndexConversion,
-  VerifierIndexConversions,
-};
-
-// basic conversion functions for each field
+export { verifierIndexConversion };
 
 import { Field } from './kimchi-types.js';
 
@@ -49,11 +42,6 @@ type WasmClasses = {
     | typeof WasmFqPlonkVerifierIndex;
 };
 
-type VerifierIndexConversion = ReturnType<
-  typeof verifierIndexConversionPerField
->;
-type VerifierIndexConversions = ReturnType<typeof verifierIndexConversion>;
-
 function verifierIndexConversion(wasm: wasm, core: ConversionCores) {
   const fp = verifierIndexConversionPerField(core.fp, {
     Domain: wasm.WasmFpDomain,
@@ -67,7 +55,6 @@ function verifierIndexConversion(wasm: wasm, core: ConversionCores) {
     Shifts: wasm.WasmFqShifts,
     VerifierIndex: wasm.WasmFqPlonkVerifierIndex,
   });
-
   return { fp, fq };
 }
